Fix MODCAT-198
