--- src/main/java/org/folio/marccat/business/cataloguing/authority/AuthorityCataloguingSourceTag.java
@@ -1,14 +1,0 @@
-package org.folio.marccat.business.cataloguing.authority;
-
-import org.folio.marccat.business.cataloguing.common.CataloguingSourceTag;
-
-public class AuthorityCataloguingSourceTag extends CataloguingSourceTag {
-
-
-  public AuthorityCataloguingSourceTag() {
-    super();
-    setHeaderField(new AuthorityHeaderFieldHelper());
-    setHeaderType((short) 1);
-  }
-
-}
