--- src/main/java/org/folio/marccat/business/cataloguing/authority/AuthorityDateOfLastTransactionTag.java
@@ -1,16 +1,0 @@
-package org.folio.marccat.business.cataloguing.authority;
-
-import org.folio.marccat.business.cataloguing.common.DateOfLastTransactionTag;
-
-
-public class AuthorityDateOfLastTransactionTag
-  extends DateOfLastTransactionTag {
-
-
-  public AuthorityDateOfLastTransactionTag() {
-    super();
-    setHeaderField(new AuthorityHeaderFieldHelper());
-    setHeaderType((short) 12);
-  }
-
-}
