--- src/main/java/org/folio/marccat/business/cataloguing/authority/AuthorityGeographicAreaTag.java
@@ -1,14 +1,0 @@
-package org.folio.marccat.business.cataloguing.authority;
-
-import org.folio.marccat.business.cataloguing.common.GeographicAreaTag;
-
-public class AuthorityGeographicAreaTag extends GeographicAreaTag {
-
-
-  public AuthorityGeographicAreaTag() {
-    super();
-    setHeaderField(new AuthorityHeaderFieldHelper());
-    setHeaderType((short) 3);
-  }
-
-}
