--- src/main/java/org/folio/marccat/business/cataloguing/authority/HasHeadingType.java
@@ -1,8 +1,0 @@
-package org.folio.marccat.business.cataloguing.authority;
-
-/**
- * Implemented by Tags that support associated headingType's and
- * the ability to change them. (e.g. Authority headings and reference tags)
- */
-public interface HasHeadingType {
-}
