--- src/main/java/org/folio/marccat/business/cataloguing/authority/SeeAlsoReferenceTag.java
@@ -1,56 +1,0 @@
-package org.folio.marccat.business.cataloguing.authority;
-
-import org.folio.marccat.dao.persistence.REF;
-import org.folio.marccat.dao.persistence.ReferenceType;
-import org.folio.marccat.shared.CorrelationValues;
-
-public class SeeAlsoReferenceTag extends SeeSeeAlsoReference {
-  private REF dualReference = null;
-
-  private short dualReferenceIndicator = 0;
-
-
-  public SeeAlsoReferenceTag() {
-    super();
-  }
-
-  @Override
-  public boolean correlationChangeAffectsKey(CorrelationValues v) {
-    if (!super.correlationChangeAffectsKey(v)) {
-      return !ReferenceType.isSeeAlsoFrom(
-        v.getValue(getRefTypeCorrelationPosition()));
-    } else {
-      return false;
-    }
-  }
-
-
-  public REF getDualReference() {
-    return dualReference;
-  }
-
-
-  public void setDualReference(REF ref) {
-    dualReference = ref;
-  }
-
-
-  @Override
-  public short getDualReferenceIndicator() {
-    return dualReferenceIndicator;
-  }
-
-
-  public void setDualReferenceIndicator(short s) {
-    dualReferenceIndicator = s;
-  }
-
-  /* (non-Javadoc)
-   * @see SeeSeeAlsoReference#getHasDualIndicator()
-   */
-  @Override
-  public boolean isHasDualIndicator() {
-    return true;
-  }
-
-}
