+++ src/main/java/org/folio/marccat/dao/DAOBibliographicNote.java
@@ -5,9 +5,9 @@ import org.folio.marccat.business.common.Persistence;
 
 public class DAOBibliographicNote extends AbstractDAO {
 
-  @Override
+ /* @Override
   public void delete(Persistence p) {
     super.delete(p);
-  }
+  }*/
 
 }
