+++ src/main/java/org/folio/marccat/integration/search/Parser.java
@@ -1,5 +1,6 @@
 package org.folio.marccat.integration.search;
 
+import net.sf.hibernate.HibernateException;
 import net.sf.hibernate.Session;
 import org.folio.marccat.business.common.View;
 import org.folio.marccat.config.log.Log;
@@ -95,7 +96,7 @@ public class Parser {
    * @return the parsed string.
    * @throws CclParserException in case of parsing failure.
    */
-  public String parse(final String ccl, final int firstRecord, final int lastRecord, final String[] attributes, String[] directions) throws CclParserException {
+  public String parse(final String ccl, final int firstRecord, final int lastRecord, final String[] attributes, String[] directions) throws CclParserException, HibernateException {
     final Tokenizer tokenizer = new Tokenizer().tokenize(ccl);
     final ExpressionNode n = parse(tokenizer.getTokens());
     final int limitSize = (lastRecord - firstRecord) + 1;
@@ -123,7 +124,7 @@ public class Parser {
    * @return the parsed string.
    * @throws CclParserException in case of parsing failure.
    */
-  public String parseAndCount(String ccl, final String[] attributes, String[] directions) throws CclParserException {
+  public String parseAndCount(String ccl, final String[] attributes, String[] directions) throws CclParserException, HibernateException {
     final Tokenizer tokenizer = new Tokenizer();
     tokenizer.tokenize(ccl);
     final ExpressionNode n = parse(tokenizer.getTokens());
@@ -144,7 +145,7 @@ public class Parser {
    * @return the expression node..
    * @throws CclParserException in case of parsing failure.
    */
-  public ExpressionNode parse(final List<Token> tokens) throws CclParserException {
+  public ExpressionNode parse(final List<Token> tokens) throws CclParserException, HibernateException {
     this.tokens = new LinkedList<>(tokens);
     this.lookahead = this.tokens.getFirst();
     ExpressionNode expr = null;
@@ -174,7 +175,7 @@ public class Parser {
    * @return the expression node
    * @throws CclParserException the ccl parser exception
    */
-  private ExpressionNode searchGroup(ExpressionNode expr) throws CclParserException {
+  private ExpressionNode searchGroup(ExpressionNode expr) throws CclParserException, HibernateException {
     if (lookahead.token == Tokenizer.TokenType.LP) {
       nextToken();
       while (lookahead.token != Tokenizer.TokenType.RP && lookahead.token != Tokenizer.TokenType.EOL) {
@@ -203,14 +204,14 @@ public class Parser {
    * @return the expression node
    * @throws CclParserException the ccl parser exception
    */
-  private ExpressionNode searchExpression() throws CclParserException {
+  private ExpressionNode searchExpression() throws CclParserException, HibernateException {
     final TermExpressionNode expr = new TermExpressionNode(session, locale, mainLibraryId, searchingView);
     if (lookahead.token == Tokenizer.TokenType.WORD) {
       if (lookahead.sequence.length() <= 3) {
         IndexList i;
         try {
           i = dao.getIndexByLocalAbbreviation(session, lookahead.sequence, locale);
-        } catch (DataAccessException e) {
+        } catch (HibernateException e) {
           throw new ModMarccatException(e);
         }
         if (i != null) {
@@ -242,7 +243,7 @@ public class Parser {
    *
    * @return the default index
    */
-  private IndexList getDefaultIndex() {
+  private IndexList getDefaultIndex() throws HibernateException {
     if (defaultIndex == null) {
       try {
         defaultIndex = dao.getIndexByLocalAbbreviation(session, "AW", Locale.ENGLISH);
